+++ src/components/search-form/search-filters.test.js
@@ -40,18 +40,6 @@ const renderSearchFilters = (props = {}) => render(
           labelId: 'Book-label',
           value: 'book',
         }],
-      }, {
-        type: FILTER_TYPES.SELECTION,
-        label: 'Content-type',
-        name: 'contentType',
-        defaultValue: 'some',
-        options: [{
-          labelId: 'Some-label',
-          value: 'some',
-        }, {
-          labelId: 'Journal-label',
-          value: 'journal',
-        }],
       }]}
       onUpdate={mockOnUpdate}
       searchType="titles"
@@ -73,7 +61,6 @@ describe('Given SearchFilters', () => {
 
     expect(getByText('Sort-label')).toBeDefined();
     expect(getByText('Publication-type')).toBeDefined();
-    expect(getByText('Content-type')).toBeDefined();
   });
 
   describe('when editing checkbox filter value', () => {
@@ -93,18 +80,6 @@ describe('Given SearchFilters', () => {
       fireEvent.change(getByRole('combobox'), { target: { value: 'book' } });
 
       expect(mockOnUpdate).toHaveBeenCalledWith(expect.objectContaining({ publicationType: 'book' }));
-    });
-  });
-
-  describe('when editing selection filter value', () => {
-    it('should call onUpdate', () => {
-      const { getByRole, getByText } = renderSearchFilters();
-
-      fireEvent.click(getByText('Content-type'));
-      fireEvent.click(getByText('stripes-components.selection.controlLabel'));
-      fireEvent.click(getByRole('option', { name : 'Journal-label' }));
-
-      expect(mockOnUpdate).toHaveBeenCalledWith(expect.objectContaining({ contentType: 'journal' }));
     });
   });