+++ src/constants/filterConfigs.js
@@ -40,7 +40,6 @@ export const selectionStatusDefaultFilterOption = selectionStatusFilterOptions.A
 
 export const FILTER_TYPES = {
   SELECT: 'select',
-  SELECTION: 'selection',
   CHECKBOX: 'checkbox',
 };
 
@@ -98,7 +97,7 @@ export const titleSortFilterConfig = {
 };
 
 export const contentTypeFilterConfig = {
-  type: FILTER_TYPES.SELECTION,
+  type: FILTER_TYPES.SELECT,
   name: 'type',
   label: <FormattedMessage id="ui-eholdings.package.contentType" />,
   defaultValue: 'all',